refactor: move CardDropHandlers to utils directory
