+++ src/components/Breach.jsx
@@ -3,7 +3,7 @@ import "./Breach.css";
 import { BASE_URL, BreachType } from "../data/constants"
 import CardDropZone from "./CardDropZone";
 import DraggableCard from "./DraggableCard";
-import { HandleCardDropOntoPile } from "./CardDropHandlers";
+import { HandleCardDropOntoPile } from "../utils/CardDropHandlers";
 import { CardLocations } from "../data/constants";
